Enhance documentation and refactor model imports
Added a new documentation page for socket requests and updated the README to include a link to it. Refactored model imports in various files to point to the new structured paths for Player and Gift models. Removed the old Gift model file and created new structured models for Gift, Creature, Egg, and Item. Updated related routes and event handlers to reflect these changes, ensuring consistency across the codebase.
