+++ src/models/gift/items.model.ts
@@ -1,0 +1,11 @@
+/**
+ * Represents an item with an identifier and a count.
+ *
+ * @interface IItems
+ * @property {string} id - The unique identifier for the item.
+ * @property {number} count - The count of the item.
+ */
+export interface IItems {
+  id: string;
+  count: number;
+}
