+++ src/models/gts/gtsIBlacklist.ts
