+++ src/websocket/events/friend/friendRequest.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 import createEventHandler from '@logic/createEventHandler';
-import { Player } from '@root/src/models/player';
+import { Player } from '@root/src/models/player/player.model';
 import { server } from '@root/src';
 
 /**
