+++ src/websocket/events/index.ts
@@ -1,21 +1,9 @@
 import { EventHandlers } from '@src/types';
 import errorHandler from './error';
-import playerCreateHandler from './player/playerCreate';
-import playerDeleteHandler from './player/playerDelete';
-import giftListHandler from './gift/giftList';
-import gtsAddHandler from './gts/gtsAdd';
-import gtsTradeHandler from './gts/gtsTrade';
-import gtsAllListHandler from './gts/gtsAllList';
-import friendRequestHandler from './friend/friendRequest';
-import friendAcceptHandler from './friend/friendAccept';
-import friendDeclineHandler from './friend/friendDecline';
-import friendRemoveHandler from './friend/friendRemove';
-import friendListHandler from './friend/friendList';
-import friendPendingHandler from './friend/friendPending';
-import giftClaimHandler from './gift/giftClaim';
-import giftClaimByCodeHandler from './gift/giftClaimByCode';
-import giftClaimByIdHandler from './gift/giftClaimById';
-import playerUpdateHandler from './player/playerUpdate';
+import { playerEvents } from './player';
+import { friendEvents } from './friend';
+import { giftEvents } from './gift';
+import { gtsEvents } from './gts';
 
 /**
  * An object containing event handlers for various websocket events.
@@ -24,22 +12,10 @@ import playerUpdateHandler from './player/playerUpdate';
  */
 const events: EventHandlers = {
   error: errorHandler,
-  playerCreate: playerCreateHandler,
-  playerDelete: playerDeleteHandler,
-  playerUpdate: playerUpdateHandler,
-  giftClaim: giftClaimHandler,
-  giftList: giftListHandler,
-  giftClaimByCode: giftClaimByCodeHandler,
-  giftClaimById: giftClaimByIdHandler,
-  gtsAdd: gtsAddHandler,
-  gtsTrade: gtsTradeHandler,
-  gtsAllList: gtsAllListHandler,
-  friendRequest: friendRequestHandler,
-  friendAccept: friendAcceptHandler,
-  friendDecline: friendDeclineHandler,
-  friendRemove: friendRemoveHandler,
-  friendList: friendListHandler,
-  friendPending: friendPendingHandler,
+  ...playerEvents,
+  ...friendEvents,
+  ...giftEvents,
+  ...gtsEvents,
 };
 
 export default events;
